// src/app/interfaces/branch.ts
export interface branch {
  CODEMPRESA: number;
  CODFILIAL: number;
  NOMEFANTASIA: string;
  RAZAOSOCIAL: string;
  CGC: string;
  INSCRICAOESTADUAL: string;
  TELEFONE: string;
  FAX: string;
  EMAIL: string;
  CONTATO: string;
  RUA: string;
  NUMERO: string;
  COMPLEMENTO: string | null;
  BAIRRO: string;
  CIDADE: string;
  ESTADO: string;
  CEP: string;
  HOMEPAGE: string;
  INATIVO: string | null;
}